backend/requirements.txt
blinker==1.9.0
click==8.1.8
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
pluggy==1.6.0
psycopg2-binary==2.9.12
Pygments==2.20.0
pytest==8.4.2
tomli==2.4.1
typing_extensions==4.15.0
Werkzeug==3.1.8
zipp==3.23.1
